work/NHDVP/requirement.txt
Django==2.2.11
mysqlclient
pymysql
djangorestframework
sqlacodegen
SQLAlchemy